Rename actions/items.js to actions/index.js
