+++ backend/utils/tmiClient.js
@@ -2,7 +2,7 @@ const tmi = require('tmi.js');
 const chooseChannel = require('./tmiConfig');
 
 const client = new tmi.client(chooseChannel([
-  'distortion2',
+  'imaqtpie',
 ]));
 
 client.connect();
