+++ src/containers/TwitchChannels/twitchChannels.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import { getChannels, setChannel } from '../../actions/items';
+import { getChannels, setChannel } from '../../actions/index';
 
 class ChannelDropdown extends Component {
   componentWillMount() {
