+++ src/reducers/reducerGetMessage.js
@@ -1,4 +1,4 @@
-import { GET_MESSAGE } from '../actions/items';
+import { GET_MESSAGE } from '../actions/index';
 
 export default function (state = [], action) {
   switch (action.type) {
